# openapi.yaml
openapi: '3.0.3'
info:
  title: Power to the People - Solar CRM API
  description: >-
    REST API for the Power to the People solar CRM platform.
    Provides endpoints for solar data queries, lead management,
    payment webhooks, referral tracking, and SMT data integration.
  version: '1.0.0'
  contact:
    email: justin@agntc.tech
servers:
  - url: https://us-central1-power-to-the-people-vpp.cloudfunctions.net
    description: Firebase Cloud Functions (us-central1)
security:
  - BearerAuth: []
paths:
...
  /solarEquipment:
    get:
      operationId: solarEquipment
      summary: Queries the solar equipment database with optional compliance and manufacturer filters
      description: Queries the solar equipment database with optional compliance and manufacturer filters
      tags:
        - solarDataApi
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: manufacturer
          in: query
          required: false
          schema:
            type: string
        - name: feoc_compliant
          in: query
          required: false
          schema:
            type: string
        - name: domestic_content
          in: query
          required: false
          schema:
            type: string
        - name: tariff_safe
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      type: boolean
                    count:
                      type: number
                    data:
                      type: array
                      items:
                        type: string
        '401':
          description: Unauthorized - invalid or missing API key
        '403':
          description: Forbidden - insufficient permissions or scope
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
  /solarUtilities:
    get:
      operationId: solarUtilities
      summary: Queries utility rate data by state, ZIP code, or utility name
      description: Queries utility rate data by state, ZIP code, or utility name
      tags:
        - solarDataApi
      security:
        - BearerAuth: []
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: zip
          in: query
          required: false
          schema:
            type: string
        - name: utility_name
          in: query
          required: false
          schema:
            type: string
        - name: has_net_metering
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      type: boolean
                    count:
                      type: number
                    data:
                      type: array
                      items:
                        type: string
        '401':
          description: Unauthorized - invalid or missing API key
        '403':
          description: Forbidden - insufficient permissions or scope
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
  /solarIncentives:
    get:
      operationId: solarIncentives
      summary: Queries active, expired, or upcoming solar incentive programs by state and sector
      description: Queries active, expired, or upcoming solar incentive programs by state and sector
      tags:
        - solarDataApi
      security:
        - BearerAuth: []
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: sector
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      type: boolean
                    count:
                      type: number
                    data:
                      type: array
                      items:
                        type: string
        '401':
          description: Unauthorized - invalid or missing API key
        '403':
          description: Forbidden - insufficient permissions or scope
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
  /solarPermits:
    get:
      operationId: solarPermits
      summary: Queries solar permit requirements by state, jurisdiction, or county
      description: Queries solar permit requirements by state, jurisdiction, or county
      tags:
        - solarDataApi
      security:
        - BearerAuth: []
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: jurisdiction
          in: query
          required: false
          schema:
            type: string
        - name: county
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      type: boolean
                    count:
                      type: number
                    data:
                      type: array
                      items:
                        type: string
        '401':
          description: Unauthorized - invalid or missing API key
        '403':
          description: Forbidden - insufficient permissions or scope
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
  /solarComplianceCheck:
    post:
      operationId: solarComplianceCheck
      summary: Runs a compound compliance report across equipment FEOC/domestic content, permits, and incentives
      description: Runs a compound compliance report across equipment FEOC/domestic content, permits, and incentives
      tags:
        - solarDataApi
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - equipment_ids
              properties:
                equipment_ids:
                  type: array
                  items:
                    type: string
                jurisdiction:
                  type: string
                state:
                  type: string
                project_type:
                  type: string
                  enum:
                    - "residential"
                    - "commercial"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      type: boolean
                    compliance_summary:
                      type: object
                    equipment:
                      type: array
                      items:
                        type: string
                    permit_requirements:
                      type: string
                    incentives:
                      type: array
                      items:
                        type: string
        '401':
          description: Unauthorized - invalid or missing API key
        '403':
          description: Forbidden - insufficient permissions or scope
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
  /solarEstimate:
    post:
      operationId: solarEstimate
      summary: Generates a full solar estimate including cost, production, incentives, equipment, permits, and financing
      description: Generates a full solar estimate including cost, production, incentives, equipment, permits, and financing
      tags:
        - solarDataApi
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - state
                - system_size_kw
              properties:
                state:
                  type: string
                zip:
                  type: string
                jurisdiction:
                  type: string
                system_size_kw:
                  type: number
                monthly_bill:
                  type: number
                project_type:
                  type: string
                  enum:
                    - "residential"
                    - "commercial"
                equipment_preferences:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      type: boolean
                    estimate:
                      type: object
                    utility_rates:
                      type: array
                      items:
                        type: string
                    recommended_equipment:
                      type: array
                      items:
                        type: string
                    incentives:
                      type: array
                      items:
                        type: string
                    permit_requirements:
                      type: string
                    financing_options:
                      type: array
                      items:
                        type: string
        '401':
          description: Unauthorized - invalid or missing API key
        '403':
          description: Forbidden - insufficient permissions or scope
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: >-
        API key authentication. Pass your API key as a Bearer token
        in the Authorization header: "Authorization: Bearer pk_live_..."
...